requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
asgiref==3.5.0
async-timeout==5.0.1
attrs==25.4.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.9.0
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dateparser==1.3.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.3.0
fastapi==0.115.12
fasteners==0.14.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
frozenlist==1.8.0
future==0.18.2
gbinder-python==1.1.2
greenlet==1.1.2
gunicorn==25.0.3
gyp==0.1
h11==0.13.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.3
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
multidict==6.7.1
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
packaging==26.0
pandas==2.3.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
propcache==0.4.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodome==3.23.0
pycups==2.0.1
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-binance==1.0.34
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
pytz==2025.2
pyxdg==0.27
PyYAML==5.4.1
regex==2026.1.15
reportlab==3.6.8
requests==2.25.1
scikit-learn==1.7.2
scipy==1.15.3
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==1.4.31
starlette==0.46.2
systemd-python==234
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.3
tzlocal==5.3.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.15.0
variety==0.8.5
wadllib==1.3.6
webencodings==0.5.1
websockets==16.0
Werkzeug==3.1.4
wsproto==1.0.0
xdg==5
xkit==0.0.0
yarl==1.22.0
zipp==1.0.0
